Add files via upload
This repository hosts the experimental measurements for the occupancy detection tasks. It includes a clear description of the data files.

Description of the data columns(units etc).
The scripts to reproduce exploratory figures.
The scripts for model training.
The commands for model testing.
Also note that when training and testing the models you have to use the seed command to ensure reproducibility.
There may be small variations in the reported accuracy.